+++ pages/test.jsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import ShowTodo from './ShowTodo'
+
+function test() {
+  return (
+    <div>
+      <ShowTodo />
+    </div>
+  )
+}
+
+export default test
